refactor: improve code quanity
- rename LogHttp to HttpLogger
