+++ tests/http_logger.rs
